Prevent "new" items from deconstructing in the Deconstructor
That I hate the Deconstructor is no news for many. But in combination with TEAM I slipped a loophole for too long.
I don't like such "solutions" though, so I will think about a consistent way of the whole deconstructing topic this year. I lean towards the idea that ammo, consumable stuff should not be deconstructible. Only full devices... if at all.
Overall the EGS Recycle and Deconstructor are very dangerous features for a good economy and player interaction.