Added HWS Creative Scenario to the Workshop!
Go here and subscribe:
https://steamcommunity.com/sharedfiles/filedetails/?id=2355482829
I see github is for too many people to complicated to use and they don't understand the nature of Cloning a repository (the proper way).
With some Junctions and Hardlinks I hope this works better and you get the latest version automatically always in the Workshop.
I will keep maintaining the github project because it is much better than steam in terms of revision etc.
It just means more effort for me...