Armor + Jetpack Boost rebalance Part 2
Considered bunch of feedback but I'm still not done how I want to have Armors and Boosters play a role in HWS.
My vision so far is: Amor alone are "OK" but only Booster will make them powerful, unique and specialized. Especially the Epic Boosters are your way to go in the future.
My vision is also that you feel progression in terms of Booster and Armors.
Hence, with this change I streamlined the Booster logic in a better way imo.
- Added 5 Boost Slots to Halloween Armor. Before it had 4.
- Added 4 Boost Slots to Epic Armor. Before it had 0
- Added 3 Boost Slots to Epic Medium Armor. Before it had 1.
- Increased Halloween Armor Durability to 1500 from 1000.
- Boosted Jump Height to compensate some lower base Jetpack value. Makes more sense.

With this change and the other stats the Halloween Armor is still a very unique item on HWS but it feels less frustrated to players who can't afford the armor or did not participate in the Event in the first place.
The Epic Armor is now a good follow up.
Overall it is a difficult topic, which neither Vanilla, nor Reforged Eden nor anyone took seriously so far. You have to test incredible lot use cases. Each planet has different gravity, which makes some Armors OP again and some not.
Testing consumes days for me, so please have patience for a good result at the end.