Cleanup and the compromise
- We are currently preparing a big Valentines Event, if you wonder what the image is for.
- I added more info to the PDA
- I added a CubeHalf for Waterblocks, thanks to Kaeser
- And I decided for a compromise about 2 issues in the game, the PhysicsCollider for ramps.
If set to false, your Ramps do not consume space when they are not active.
If set to true, your ramps do always consume the space even if they are not active. That means an invisible barrier is preventing your ships from landing to a wall in certain situation for example. Really depends on the design.
In the end though it prevents two glitches or exploit in combination with weapons and the player, glitch through it.
Overall, setting it to false does not make much sense, cause in the end, your ramp would glitch through the terrain or blocks for example.